move player preview into it's own file
